Rename samples -> demos
