+++ demos/appyx-interactions/desktop/src/desktopMain/kotlin/com/bumble/appyx/interactions/widgets/Widgets.kt
@@ -18,15 +18,15 @@ import com.bumble.appyx.components.spotlight.SpotlightModel
 import com.bumble.appyx.components.spotlight.operation.next
 import com.bumble.appyx.components.spotlight.operation.previous
 import com.bumble.appyx.components.spotlight.ui.slider.SpotlightSlider
+import com.bumble.appyx.demos.common.widget.CalendarWidget
+import com.bumble.appyx.demos.common.widget.TimerWidget
+import com.bumble.appyx.demos.common.widget.WeatherWidget
 import com.bumble.appyx.interactions.Events
 import com.bumble.appyx.interactions.core.AppyxInteractionsContainer
 import com.bumble.appyx.interactions.core.ui.gesture.GestureSettleConfig
 import com.bumble.appyx.interactions.core.ui.helper.AppyxComponentSetup
 import com.bumble.appyx.interactions.core.ui.output.ElementUiModel
 import com.bumble.appyx.interactions.widgets.ui.WidgetsStack3D
-import com.bumble.appyx.samples.common.widget.CalendarWidget
-import com.bumble.appyx.samples.common.widget.TimerWidget
-import com.bumble.appyx.samples.common.widget.WeatherWidget
 import kotlinx.coroutines.flow.Flow
 
 enum class WidgetsType {
