+++ demos/common/src/androidMain/kotlin/com/bumble/appyx/demos/common/profile/ProfileCardNode.kt
@@ -1,4 +1,4 @@
-package com.bumble.appyx.samples.common.profile
+package com.bumble.appyx.demos.common.profile
 
 import androidx.compose.foundation.layout.BoxWithConstraints
 import androidx.compose.foundation.layout.Column
